requirements.txt
-i https://pypi.org/simple
boto3==1.9.217
botocore==1.12.217
certifi==2019.6.16
chardet==3.0.4
defusedxml==0.6.0
django-allauth==0.39.1
django-bootstrap4==0.0.8
django-debug-toolbar==2.0
django-environ==0.4.5
django-storages==1.7.1
django==2.2.4
docutils==0.15.2
idna==2.8
jmespath==0.9.4
oauthlib==3.1.0
pillow==6.1.0
psycopg2-binary==2.8.3
python-dateutil==2.8.0 ; python_version >= '2.7'
python3-openid==3.1.0
pytz==2019.2
requests-oauthlib==1.2.0
requests==2.22.0
s3transfer==0.2.1
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3 ; python_version >= '3.4'
